refactor(board): board read model name and add builder annotation
