+++ src/main/java/me/nettee/board/application/model/BoardQueryModels.java
@@ -1,14 +1,16 @@
 package me.nettee.board.application.model;
 
+import lombok.Builder;
 import me.nettee.board.application.domain.type.BoardStatus;
 
 import java.time.Instant;
 
-public final class BoardQueryModel {
+public final class BoardQueryModels {
 
-    private BoardQueryModel() {
+    private BoardQueryModels() {
     }
 
+    @Builder
     public record BoardDetail(
             Long id,
             String title,
@@ -19,6 +21,7 @@ public final class BoardQueryModel {
     ) {
     }
 
+    @Builder
     public record BoardSummary(
             Long id,
             String title,
