+++ src/main/java/me/nettee/board/application/port/BoardQueryPort.java
@@ -3,8 +3,8 @@ package me.nettee.board.application.port;
 import java.util.Optional;
 import java.util.Set;
 import me.nettee.board.application.domain.type.BoardStatus;
-import me.nettee.board.application.model.BoardQueryModel.BoardDetail;
-import me.nettee.board.application.model.BoardQueryModel.BoardSummary;
+import me.nettee.board.application.model.BoardQueryModels.BoardDetail;
+import me.nettee.board.application.model.BoardQueryModels.BoardSummary;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
